+++ src/lib/radar/trail/cartesian.rs
@@ -3,8 +3,8 @@ use ndarray::{Array2, ArrayBase, Dim, OwnedRepr};
 #[allow(dead_code)]
 #[derive(Clone, Copy)]
 pub struct Point {
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
 }
 
 #[derive(Clone, Copy)]
@@ -27,12 +27,12 @@ impl PolarToCartesianLookup {
         let mut xyi = Vec::with_capacity(spokes_per_revolution * spoke_len);
         for arc in 0..spokes_per_revolution {
             let sine =
-                (arc as f32 * 2.0 * std::f32::consts::PI / spokes_per_revolution as f32).sin();
+                (arc as f64 * 2.0 * std::f64::consts::PI / spokes_per_revolution as f64).sin();
             let cosine =
-                (arc as f32 * 2.0 * std::f32::consts::PI / spokes_per_revolution as f32).cos();
+                (arc as f64 * 2.0 * std::f64::consts::PI / spokes_per_revolution as f64).cos();
             for radius in 0..spoke_len {
-                let x = radius as f32 * cosine;
-                let y = radius as f32 * sine;
+                let x = radius as f64 * cosine;
+                let y = radius as f64 * sine;
                 xy.push(Point { x, y });
                 xyi.push(PointInt {
                     x: x as i16,
